fixed possible overflow on InventoryMenu::onTakeStuff()

+++ Game/ui/inventorymenu.cpp
@@ -337,7 +337,7 @@ InventoryMenu::PageLocal &InventoryMenu::activePageSel() {
   }
 
 void InventoryMenu::onTakeStuff() { 
-  uint8_t itemCount = 0;
+  uint32_t itemCount = 0;
   auto& page = activePage();
   auto& sel = activePageSel();
   if(sel.sel >= page.size())
@@ -347,7 +347,7 @@ void InventoryMenu::onTakeStuff() {
     ++takeCount;
     itemCount = pow(10,takeCount / 10);
     if(r.count() <= itemCount) {
-      itemCount = r.count();
+      itemCount = uint32_t(r.count());
       takeCount = 0;
     }
   }
